@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -31,6 +31,7 @@ dependencies:
   built_value: 8.1.3
   cached_network_image: 3.1.0+1
   cloud_firestore: 2.5.4
+  cloud_firestore_platform_interface: 5.6.0
   cloud_firestore_web: 2.6.8
   cloud_functions: 3.1.1
   collection: 1.15.0
@@ -39,9 +40,13 @@ dependencies:
   equatable: 2.0.3
   file_picker: 4.5.1
   firebase_auth: 3.2.0
+  firebase_auth_platform_interface: 6.3.1
+  firebase_auth_web: 3.3.7
   firebase_core: 1.12.0
+  firebase_core_platform_interface: 4.4.1
   firebase_core_web: 1.5.4
   firebase_storage: 10.1.0
+  firebase_storage_platform_interface: 4.1.11
   firebase_storage_web: 3.2.8
   floating_bottom_navigation_bar: 1.5.2
   flutter_datetime_picker: 1.5.1
@@ -58,7 +63,8 @@ dependencies:
   json_serializable: 6.1.3
   mime_type: 1.0.0
   page_transition: 2.0.4
-  rxdart: 0.26.0
+  plugin_platform_interface: 2.1.2
+  rxdart: 0.27.4
   share_plus: 2.1.4
   shared_preferences: 2.0.11
   sign_in_with_apple:
@@ -76,6 +82,7 @@ dependencies:
   cupertino_icons: ^1.0.0
 
 dependency_overrides:
+  rxdart: 0.27.4
   stripe_android: 2.4.0
 
 dev_dependencies:
