@@ -4,63 +4,133 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 abstract class FlutterFlowTheme {
-  static FlutterFlowTheme of(BuildContext context) => LightModeTheme();
+  static FlutterFlowTheme of(BuildContext context) {
+    return LightModeTheme();
+  }
 
-  Color primaryColor;
-  Color secondaryColor;
-  Color tertiaryColor;
-  Color alternate;
-  Color primaryBackground;
-  Color secondaryBackground;
-  Color primaryText;
-  Color secondaryText;
+  late Color primaryColor;
+  late Color secondaryColor;
+  late Color tertiaryColor;
+  late Color alternate;
+  late Color primaryBackground;
+  late Color secondaryBackground;
+  late Color primaryText;
+  late Color secondaryText;
 
-  Color salmonPink;
-  Color pinkLace;
-  Color magicMint;
-  Color black;
-  Color magicMint2;
-  Color cultured;
-  Color softPink;
-  Color purplePastel;
-  Color greenPastel;
+  late Color salmonPink;
+  late Color pinkLace;
+  late Color magicMint;
+  late Color black;
+  late Color magicMint2;
+  late Color cultured;
+  late Color softPink;
+  late Color purplePastel;
+  late Color greenPastel;
 
+  String get title1Family => typography.title1Family;
+  TextStyle get title1 => typography.title1;
+  String get title2Family => typography.title2Family;
+  TextStyle get title2 => typography.title2;
+  String get title3Family => typography.title3Family;
+  TextStyle get title3 => typography.title3;
+  String get subtitle1Family => typography.subtitle1Family;
+  TextStyle get subtitle1 => typography.subtitle1;
+  String get subtitle2Family => typography.subtitle2Family;
+  TextStyle get subtitle2 => typography.subtitle2;
+  String get bodyText1Family => typography.bodyText1Family;
+  TextStyle get bodyText1 => typography.bodyText1;
+  String get bodyText2Family => typography.bodyText2Family;
+  TextStyle get bodyText2 => typography.bodyText2;
+
+  Typography get typography => ThemeTypography(this);
+}
+
+class LightModeTheme extends FlutterFlowTheme {
+  late Color primaryColor = const Color(0xFF3474E0);
+  late Color secondaryColor = const Color(0xFFEE8B60);
+  late Color tertiaryColor = const Color(0xFFFFFFFF);
+  late Color alternate = const Color(0x00000000);
+  late Color primaryBackground = const Color(0x00000000);
+  late Color secondaryBackground = const Color(0x00000000);
+  late Color primaryText = const Color(0x00000000);
+  late Color secondaryText = const Color(0x00000000);
+
+  late Color salmonPink = Color(0xFFFF999D);
+  late Color pinkLace = Color(0xFFFFDCF4);
+  late Color magicMint = Color(0xFF9DF7E5);
+  late Color black = Color(0xFF000000);
+  late Color magicMint2 = Color(0xFFB9F1D9);
+  late Color cultured = Color(0xFFF4F4F4);
+  late Color softPink = Color(0xFFFADADD);
+  late Color purplePastel = Color(0xFFFDBEEB);
+  late Color greenPastel = Color(0xFFBEFDE5);
+}
+
+abstract class Typography {
+  String get title1Family;
+  TextStyle get title1;
+  String get title2Family;
+  TextStyle get title2;
+  String get title3Family;
+  TextStyle get title3;
+  String get subtitle1Family;
+  TextStyle get subtitle1;
+  String get subtitle2Family;
+  TextStyle get subtitle2;
+  String get bodyText1Family;
+  TextStyle get bodyText1;
+  String get bodyText2Family;
+  TextStyle get bodyText2;
+}
+
+class ThemeTypography extends Typography {
+  ThemeTypography(this.theme);
+
+  final FlutterFlowTheme theme;
+
+  String get title1Family => 'Poppins';
   TextStyle get title1 => GoogleFonts.getFont(
         'Poppins',
         color: Color(0xFF303030),
         fontWeight: FontWeight.w900,
         fontSize: 38,
       );
+  String get title2Family => 'Poppins';
   TextStyle get title2 => GoogleFonts.getFont(
         'Poppins',
         color: Color(0xFF303030),
         fontWeight: FontWeight.w500,
         fontSize: 22,
       );
+  String get title3Family => 'Poppins';
   TextStyle get title3 => GoogleFonts.getFont(
         'Poppins',
         color: Color(0xFF303030),
         fontWeight: FontWeight.w500,
         fontSize: 20,
       );
+  String get subtitle1Family => 'Poppins';
   TextStyle get subtitle1 => GoogleFonts.getFont(
         'Poppins',
-        color: black,
+        color: theme.black,
         fontWeight: FontWeight.bold,
         fontSize: 18,
       );
+  String get subtitle2Family => 'Poppins';
   TextStyle get subtitle2 => GoogleFonts.getFont(
         'Poppins',
         color: Color(0xFF616161),
         fontWeight: FontWeight.normal,
         fontSize: 16,
       );
+  String get bodyText1Family => 'Poppins';
   TextStyle get bodyText1 => GoogleFonts.getFont(
         'Poppins',
         color: Color(0xFF303030),
         fontWeight: FontWeight.normal,
         fontSize: 14,
       );
+  String get bodyText2Family => 'Poppins';
   TextStyle get bodyText2 => GoogleFonts.getFont(
         'Poppins',
         color: Color(0xFF424242),
@@ -69,43 +139,24 @@ abstract class FlutterFlowTheme {
       );
 }
 
-class LightModeTheme extends FlutterFlowTheme {
-  Color primaryColor = const Color(0xFF3474E0);
-  Color secondaryColor = const Color(0xFFEE8B60);
-  Color tertiaryColor = const Color(0xFFFFFFFF);
-  Color alternate = const Color(0x00000000);
-  Color primaryBackground = const Color(0x00000000);
-  Color secondaryBackground = const Color(0x00000000);
-  Color primaryText = const Color(0x00000000);
-  Color secondaryText = const Color(0x00000000);
-
-  Color salmonPink = Color(0xFFFF999D);
-  Color pinkLace = Color(0xFFFFDCF4);
-  Color magicMint = Color(0xFF9DF7E5);
-  Color black = Color(0xFF000000);
-  Color magicMint2 = Color(0xFFB9F1D9);
-  Color cultured = Color(0xFFF4F4F4);
-  Color softPink = Color(0xFFFADADD);
-  Color purplePastel = Color(0xFFFDBEEB);
-  Color greenPastel = Color(0xFFBEFDE5);
-}
-
 extension TextStyleHelper on TextStyle {
   TextStyle override({
-    String fontFamily,
-    Color color,
-    double fontSize,
-    FontWeight fontWeight,
-    FontStyle fontStyle,
+    String? fontFamily,
+    Color? color,
+    double? fontSize,
+    FontWeight? fontWeight,
+    double? letterSpacing,
+    FontStyle? fontStyle,
     bool useGoogleFonts = true,
-    TextDecoration decoration,
-    double lineHeight,
+    TextDecoration? decoration,
+    double? lineHeight,
   }) =>
       useGoogleFonts
           ? GoogleFonts.getFont(
-              fontFamily,
+              fontFamily!,
               color: color ?? this.color,
               fontSize: fontSize ?? this.fontSize,
+              letterSpacing: letterSpacing ?? this.letterSpacing,
               fontWeight: fontWeight ?? this.fontWeight,
               fontStyle: fontStyle ?? this.fontStyle,
               decoration: decoration,
@@ -115,6 +166,7 @@ extension TextStyleHelper on TextStyle {
               fontFamily: fontFamily,
               color: color,
               fontSize: fontSize,
+              letterSpacing: letterSpacing,
               fontWeight: fontWeight,
               fontStyle: fontStyle,
               decoration: decoration,
