@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 
 class ToggleIcon extends StatelessWidget {
   const ToggleIcon({
-    @required this.value,
-    @required this.onPressed,
-    @required this.onIcon,
-    @required this.offIcon,
+    required this.value,
+    required this.onPressed,
+    required this.onIcon,
+    required this.offIcon,
   });
 
   final bool value;
