@@ -8,7 +8,7 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class LoginWidget extends StatefulWidget {
-  const LoginWidget({Key key}) : super(key: key);
+  const LoginWidget({Key? key}) : super(key: key);
 
   @override
   _LoginWidgetState createState() => _LoginWidgetState();
@@ -129,7 +129,8 @@ class _LoginWidgetState extends State<LoginWidget> {
                                                 color: Colors.transparent,
                                                 width: 1,
                                               ),
-                                              borderRadius: 34,
+                                              borderRadius:
+                                                  BorderRadius.circular(34),
                                             ),
                                           ),
                                         ],
@@ -203,7 +204,9 @@ class _LoginWidgetState extends State<LoginWidget> {
                                                                 .transparent,
                                                             width: 0,
                                                           ),
-                                                          borderRadius: 34,
+                                                          borderRadius:
+                                                              BorderRadius
+                                                                  .circular(34),
                                                         ),
                                                       ),
                                                     ),
@@ -284,7 +287,9 @@ class _LoginWidgetState extends State<LoginWidget> {
                                                       color: Colors.transparent,
                                                       width: 0,
                                                     ),
-                                                    borderRadius: 34,
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            34),
                                                   ),
                                                 ),
                                               ),
@@ -352,7 +357,8 @@ class _LoginWidgetState extends State<LoginWidget> {
                                                   color: Colors.transparent,
                                                   width: 0,
                                                 ),
-                                                borderRadius: 34,
+                                                borderRadius:
+                                                    BorderRadius.circular(34),
                                               ),
                                             ),
                                     ],
@@ -461,7 +467,8 @@ class _LoginWidgetState extends State<LoginWidget> {
                                                 color: Colors.transparent,
                                                 width: 1,
                                               ),
-                                              borderRadius: 34,
+                                              borderRadius:
+                                                  BorderRadius.circular(34),
                                             ),
                                           ),
                                         ],
@@ -535,7 +542,9 @@ class _LoginWidgetState extends State<LoginWidget> {
                                                                 .transparent,
                                                             width: 0,
                                                           ),
-                                                          borderRadius: 34,
+                                                          borderRadius:
+                                                              BorderRadius
+                                                                  .circular(34),
                                                         ),
                                                       ),
                                                     ),
@@ -616,7 +625,9 @@ class _LoginWidgetState extends State<LoginWidget> {
                                                       color: Colors.transparent,
                                                       width: 0,
                                                     ),
-                                                    borderRadius: 34,
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            34),
                                                   ),
                                                 ),
                                               ),
@@ -684,7 +695,8 @@ class _LoginWidgetState extends State<LoginWidget> {
                                                   color: Colors.transparent,
                                                   width: 0,
                                                 ),
-                                                borderRadius: 34,
+                                                borderRadius:
+                                                    BorderRadius.circular(34),
                                               ),
                                             ),
                                     ],
