@@ -10,7 +10,7 @@ import 'package:google_fonts/google_fonts.dart';
 
 class ProfileWidget extends StatefulWidget {
   const ProfileWidget({
-    Key key,
+    Key? key,
     this.state,
   }) : super(key: key);
 
@@ -21,7 +21,7 @@ class ProfileWidget extends StatefulWidget {
 }
 
 class _ProfileWidgetState extends State<ProfileWidget> {
-  TextEditingController tourNameTextFieldController;
+  TextEditingController? tourNameTextFieldController;
   final scaffoldKey = GlobalKey<ScaffoldState>();
 
   @override
@@ -176,7 +176,7 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                   child: FFButtonWidget(
                                     onPressed: () async {
                                       if (!(functions.isStringNotEmpty(
-                                          tourNameTextFieldController.text))) {
+                                          tourNameTextFieldController!.text))) {
                                         ScaffoldMessenger.of(context)
                                             .showSnackBar(
                                           SnackBar(
@@ -197,7 +197,7 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                       } else {
                                         setState(() => FFAppState()
                                                 .newTourName =
-                                            tourNameTextFieldController.text);
+                                            tourNameTextFieldController!.text);
                                       }
 
                                       context.pushNamed(
@@ -229,7 +229,7 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                         color: Colors.transparent,
                                         width: 1,
                                       ),
-                                      borderRadius: 34,
+                                      borderRadius: BorderRadius.circular(34),
                                     ),
                                   ),
                                 ),
@@ -406,7 +406,7 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                         }
                                         List<ToursRecord>
                                             listViewToursRecordList =
-                                            snapshot.data;
+                                            snapshot.data!;
                                         if (listViewToursRecordList.isEmpty) {
                                           return CreateNewTourEmptyStateWidget();
                                         }
@@ -460,8 +460,8 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                                               RegionsRecord>(
                                                             stream: RegionsRecord
                                                                 .getDocument(
-                                                                    listViewToursRecord
-                                                                        .regionID),
+                                                                    listViewToursRecord!
+                                                                        .regionID!),
                                                             builder: (context,
                                                                 snapshot) {
                                                               // Customize what your widget looks like when it's loading.
@@ -482,7 +482,8 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                                                 );
                                                               }
                                                               final rowRegionsRecord =
-                                                                  snapshot.data;
+                                                                  snapshot
+                                                                      .data!;
                                                               return Row(
                                                                 mainAxisSize:
                                                                     MainAxisSize
@@ -538,7 +539,7 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                                                                         ClipRRect(
                                                                                           borderRadius: BorderRadius.circular(28),
                                                                                           child: Image.network(
-                                                                                            rowRegionsRecord.image,
+                                                                                            rowRegionsRecord!.image!,
                                                                                             width: MediaQuery.of(context).size.width * 0.34,
                                                                                             height: MediaQuery.of(context).size.height * 0.2,
                                                                                             fit: BoxFit.cover,
@@ -554,7 +555,7 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                                                                           child: Align(
                                                                                             alignment: AlignmentDirectional(0, 0),
                                                                                             child: Text(
-                                                                                              rowRegionsRecord.name,
+                                                                                              rowRegionsRecord!.name!,
                                                                                               textAlign: TextAlign.center,
                                                                                               style: FlutterFlowTheme.of(context).bodyText1.override(
                                                                                                     fontFamily: 'Poppins',
@@ -583,7 +584,7 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                                                                           Padding(
                                                                                             padding: EdgeInsetsDirectional.fromSTEB(24, 0, 4, 0),
                                                                                             child: Text(
-                                                                                              listViewToursRecord.tourName.maybeHandleOverflow(
+                                                                                              listViewToursRecord!.tourName!.maybeHandleOverflow(
                                                                                                 maxChars: 16,
                                                                                                 replacement: '…',
                                                                                               ),
@@ -610,7 +611,7 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                                                                           Padding(
                                                                                             padding: EdgeInsetsDirectional.fromSTEB(10, 0, 0, 0),
                                                                                             child: Text(
-                                                                                              listViewToursRecord.passengers.toString().maybeHandleOverflow(
+                                                                                              listViewToursRecord!.passengers!.toString().maybeHandleOverflow(
                                                                                                     maxChars: 25,
                                                                                                     replacement: '…',
                                                                                                   ),
@@ -638,7 +639,7 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                                                                           Padding(
                                                                                             padding: EdgeInsetsDirectional.fromSTEB(10, 0, 0, 0),
                                                                                             child: Text(
-                                                                                              dateTimeFormat('MMMMEEEEd', listViewToursRecord.tourDate),
+                                                                                              dateTimeFormat('MMMMEEEEd', listViewToursRecord!.tourDate!),
                                                                                               style: FlutterFlowTheme.of(context).bodyText1.override(
                                                                                                     fontFamily: 'Poppins',
                                                                                                     fontSize: 12,
@@ -662,7 +663,7 @@ class _ProfileWidgetState extends State<ProfileWidget> {
                                                                                           Padding(
                                                                                             padding: EdgeInsetsDirectional.fromSTEB(10, 0, 4, 0),
                                                                                             child: Text(
-                                                                                              listViewToursRecord.pickupAddress.maybeHandleOverflow(
+                                                                                              listViewToursRecord!.pickupAddress!.maybeHandleOverflow(
                                                                                                 maxChars: 20,
                                                                                                 replacement: '…',
                                                                                               ),
