@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 import 'auth_util.dart';
 
-Future<User> signInAnonymously(BuildContext context) async {
+Future<User?> signInAnonymously(BuildContext context) async {
   final signInFunc = () => FirebaseAuth.instance.signInAnonymously();
   return signInOrCreateAccount(context, signInFunc);
 }
