@@ -3,14 +3,14 @@ import 'package:flutter/material.dart';
 
 import 'auth_util.dart';
 
-Future<User> signInWithEmail(
+Future<User?> signInWithEmail(
     BuildContext context, String email, String password) async {
   final signInFunc = () => FirebaseAuth.instance
       .signInWithEmailAndPassword(email: email.trim(), password: password);
   return signInOrCreateAccount(context, signInFunc);
 }
 
-Future<User> createAccountWithEmail(
+Future<User?> createAccountWithEmail(
     BuildContext context, String email, String password) async {
   final createAccountFunc = () => FirebaseAuth.instance
       .createUserWithEmailAndPassword(email: email.trim(), password: password);
