@@ -3,7 +3,7 @@ import 'dart:typed_data';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:mime_type/mime_type.dart';
 
-Future<String> uploadData(String path, Uint8List data) async {
+Future<String?> uploadData(String path, Uint8List data) async {
   final storageRef = FirebaseStorage.instance.ref().child(path);
   final metadata = SettableMetadata(contentType: mime(path));
   final result = await storageRef.putData(data, metadata);
