@@ -10,21 +10,17 @@ abstract class RegionsRecord
     implements Built<RegionsRecord, RegionsRecordBuilder> {
   static Serializer<RegionsRecord> get serializer => _$regionsRecordSerializer;
 
-  @nullable
-  String get name;
+  String? get name;
 
-  @nullable
-  String get image;
+  String? get image;
 
-  @nullable
-  bool get isServiced;
+  bool? get isServiced;
 
-  @nullable
-  String get regionID;
+  String? get regionID;
 
-  @nullable
   @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
+  DocumentReference? get ffRef;
+  DocumentReference get reference => ffRef!;
 
   static void _initializeBuilder(RegionsRecordBuilder builder) => builder
     ..name = ''
@@ -37,11 +33,11 @@ abstract class RegionsRecord
 
   static Stream<RegionsRecord> getDocument(DocumentReference ref) => ref
       .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .map((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   static Future<RegionsRecord> getDocumentOnce(DocumentReference ref) => ref
       .get()
-      .then((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .then((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   RegionsRecord._();
   factory RegionsRecord([void Function(RegionsRecordBuilder) updates]) =
@@ -50,19 +46,25 @@ abstract class RegionsRecord
   static RegionsRecord getDocumentFromData(
           Map<String, dynamic> data, DocumentReference reference) =>
       serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+          {...mapFromFirestore(data), kDocumentReferenceField: reference})!;
 }
 
 Map<String, dynamic> createRegionsRecordData({
-  String name,
-  String image,
-  bool isServiced,
-  String regionID,
-}) =>
-    serializers.toFirestore(
-        RegionsRecord.serializer,
-        RegionsRecord((r) => r
-          ..name = name
-          ..image = image
-          ..isServiced = isServiced
-          ..regionID = regionID));
+  String? name,
+  String? image,
+  bool? isServiced,
+  String? regionID,
+}) {
+  final firestoreData = serializers.toFirestore(
+    RegionsRecord.serializer,
+    RegionsRecord(
+      (r) => r
+        ..name = name
+        ..image = image
+        ..isServiced = isServiced
+        ..regionID = regionID,
+    ),
+  );
+
+  return firestoreData;
+}
