@@ -11,55 +11,42 @@ abstract class SelectedVenuesRecord
   static Serializer<SelectedVenuesRecord> get serializer =>
       _$selectedVenuesRecordSerializer;
 
-  @nullable
-  DocumentReference get venueRef;
+  DocumentReference? get venueRef;
 
-  @nullable
-  DocumentReference get tourRef;
+  DocumentReference? get tourRef;
 
-  @nullable
   @BuiltValueField(wireName: 'added_by_uid')
-  DocumentReference get addedByUid;
+  DocumentReference? get addedByUid;
 
-  @nullable
   @BuiltValueField(wireName: 'is_lunch_venue')
-  bool get isLunchVenue;
+  bool? get isLunchVenue;
 
-  @nullable
-  double get tastingFee;
+  double? get tastingFee;
 
-  @nullable
-  DateTime get addedDate;
+  DateTime? get addedDate;
 
-  @nullable
-  String get bookingReference;
+  String? get bookingReference;
 
-  @nullable
-  DateTime get reservationTime;
+  DateTime? get reservationTime;
 
-  @nullable
   @BuiltValueField(wireName: 'region_ID')
-  DocumentReference get regionID;
+  DocumentReference? get regionID;
 
-  @nullable
   @BuiltValueField(wireName: 'is_large_group_early_seating_only_venue')
-  bool get isLargeGroupEarlySeatingOnlyVenue;
+  bool? get isLargeGroupEarlySeatingOnlyVenue;
 
-  @nullable
   @BuiltValueField(wireName: 'is_lunch_venue_only')
-  bool get isLunchVenueOnly;
+  bool? get isLunchVenueOnly;
 
-  @nullable
   @BuiltValueField(wireName: 'tasting_experience_description')
-  String get tastingExperienceDescription;
+  String? get tastingExperienceDescription;
 
-  @nullable
   @BuiltValueField(wireName: 'is_tasting_included')
-  bool get isTastingIncluded;
+  bool? get isTastingIncluded;
 
-  @nullable
   @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
+  DocumentReference? get ffRef;
+  DocumentReference get reference => ffRef!;
 
   static void _initializeBuilder(SelectedVenuesRecordBuilder builder) => builder
     ..isLunchVenue = false
@@ -75,11 +62,11 @@ abstract class SelectedVenuesRecord
 
   static Stream<SelectedVenuesRecord> getDocument(DocumentReference ref) => ref
       .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .map((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   static Future<SelectedVenuesRecord> getDocumentOnce(DocumentReference ref) =>
       ref.get().then(
-          (s) => serializers.deserializeWith(serializer, serializedData(s)));
+          (s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   SelectedVenuesRecord._();
   factory SelectedVenuesRecord(
@@ -89,38 +76,43 @@ abstract class SelectedVenuesRecord
   static SelectedVenuesRecord getDocumentFromData(
           Map<String, dynamic> data, DocumentReference reference) =>
       serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+          {...mapFromFirestore(data), kDocumentReferenceField: reference})!;
 }
 
 Map<String, dynamic> createSelectedVenuesRecordData({
-  DocumentReference venueRef,
-  DocumentReference tourRef,
-  DocumentReference addedByUid,
-  bool isLunchVenue,
-  double tastingFee,
-  DateTime addedDate,
-  String bookingReference,
-  DateTime reservationTime,
-  DocumentReference regionID,
-  bool isLargeGroupEarlySeatingOnlyVenue,
-  bool isLunchVenueOnly,
-  String tastingExperienceDescription,
-  bool isTastingIncluded,
-}) =>
-    serializers.toFirestore(
-        SelectedVenuesRecord.serializer,
-        SelectedVenuesRecord((s) => s
-          ..venueRef = venueRef
-          ..tourRef = tourRef
-          ..addedByUid = addedByUid
-          ..isLunchVenue = isLunchVenue
-          ..tastingFee = tastingFee
-          ..addedDate = addedDate
-          ..bookingReference = bookingReference
-          ..reservationTime = reservationTime
-          ..regionID = regionID
-          ..isLargeGroupEarlySeatingOnlyVenue =
-              isLargeGroupEarlySeatingOnlyVenue
-          ..isLunchVenueOnly = isLunchVenueOnly
-          ..tastingExperienceDescription = tastingExperienceDescription
-          ..isTastingIncluded = isTastingIncluded));
+  DocumentReference? venueRef,
+  DocumentReference? tourRef,
+  DocumentReference? addedByUid,
+  bool? isLunchVenue,
+  double? tastingFee,
+  DateTime? addedDate,
+  String? bookingReference,
+  DateTime? reservationTime,
+  DocumentReference? regionID,
+  bool? isLargeGroupEarlySeatingOnlyVenue,
+  bool? isLunchVenueOnly,
+  String? tastingExperienceDescription,
+  bool? isTastingIncluded,
+}) {
+  final firestoreData = serializers.toFirestore(
+    SelectedVenuesRecord.serializer,
+    SelectedVenuesRecord(
+      (s) => s
+        ..venueRef = venueRef
+        ..tourRef = tourRef
+        ..addedByUid = addedByUid
+        ..isLunchVenue = isLunchVenue
+        ..tastingFee = tastingFee
+        ..addedDate = addedDate
+        ..bookingReference = bookingReference
+        ..reservationTime = reservationTime
+        ..regionID = regionID
+        ..isLargeGroupEarlySeatingOnlyVenue = isLargeGroupEarlySeatingOnlyVenue
+        ..isLunchVenueOnly = isLunchVenueOnly
+        ..tastingExperienceDescription = tastingExperienceDescription
+        ..isTastingIncluded = isTastingIncluded,
+    ),
+  );
+
+  return firestoreData;
+}
