@@ -12,21 +12,18 @@ abstract class TastingExperiencesRecord
   static Serializer<TastingExperiencesRecord> get serializer =>
       _$tastingExperiencesRecordSerializer;
 
-  @nullable
-  String get image;
+  String? get image;
 
-  @nullable
-  String get description;
+  String? get description;
 
-  @nullable
   @BuiltValueField(wireName: 'tasting_experience_price')
-  double get tastingExperiencePrice;
+  double? get tastingExperiencePrice;
 
-  @nullable
   @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
+  DocumentReference? get ffRef;
+  DocumentReference get reference => ffRef!;
 
-  DocumentReference get parentReference => reference.parent.parent;
+  DocumentReference get parentReference => reference.parent.parent!;
 
   static void _initializeBuilder(TastingExperiencesRecordBuilder builder) =>
       builder
@@ -34,7 +31,7 @@ abstract class TastingExperiencesRecord
         ..description = ''
         ..tastingExperiencePrice = 0.0;
 
-  static Query<Map<String, dynamic>> collection([DocumentReference parent]) =>
+  static Query<Map<String, dynamic>> collection([DocumentReference? parent]) =>
       parent != null
           ? parent.collection('tasting_experiences')
           : FirebaseFirestore.instance.collectionGroup('tasting_experiences');
@@ -44,12 +41,12 @@ abstract class TastingExperiencesRecord
 
   static Stream<TastingExperiencesRecord> getDocument(DocumentReference ref) =>
       ref.snapshots().map(
-          (s) => serializers.deserializeWith(serializer, serializedData(s)));
+          (s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   static Future<TastingExperiencesRecord> getDocumentOnce(
           DocumentReference ref) =>
       ref.get().then(
-          (s) => serializers.deserializeWith(serializer, serializedData(s)));
+          (s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   TastingExperiencesRecord._();
   factory TastingExperiencesRecord(
@@ -59,17 +56,23 @@ abstract class TastingExperiencesRecord
   static TastingExperiencesRecord getDocumentFromData(
           Map<String, dynamic> data, DocumentReference reference) =>
       serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+          {...mapFromFirestore(data), kDocumentReferenceField: reference})!;
 }
 
 Map<String, dynamic> createTastingExperiencesRecordData({
-  String image,
-  String description,
-  double tastingExperiencePrice,
-}) =>
-    serializers.toFirestore(
-        TastingExperiencesRecord.serializer,
-        TastingExperiencesRecord((t) => t
-          ..image = image
-          ..description = description
-          ..tastingExperiencePrice = tastingExperiencePrice));
+  String? image,
+  String? description,
+  double? tastingExperiencePrice,
+}) {
+  final firestoreData = serializers.toFirestore(
+    TastingExperiencesRecord.serializer,
+    TastingExperiencesRecord(
+      (t) => t
+        ..image = image
+        ..description = description
+        ..tastingExperiencePrice = tastingExperiencePrice,
+    ),
+  );
+
+  return firestoreData;
+}
