@@ -11,33 +11,29 @@ abstract class TourMessagesRecord
   static Serializer<TourMessagesRecord> get serializer =>
       _$tourMessagesRecordSerializer;
 
-  @nullable
   @BuiltValueField(wireName: 'date_sent')
-  DateTime get dateSent;
+  DateTime? get dateSent;
 
-  @nullable
   @BuiltValueField(wireName: 'sender_user_reff')
-  DocumentReference get senderUserReff;
+  DocumentReference? get senderUserReff;
 
-  @nullable
   @BuiltValueField(wireName: 'message_body')
-  String get messageBody;
+  String? get messageBody;
 
-  @nullable
   @BuiltValueField(wireName: 'is_read')
-  bool get isRead;
+  bool? get isRead;
 
-  @nullable
   @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
+  DocumentReference? get ffRef;
+  DocumentReference get reference => ffRef!;
 
-  DocumentReference get parentReference => reference.parent.parent;
+  DocumentReference get parentReference => reference.parent.parent!;
 
   static void _initializeBuilder(TourMessagesRecordBuilder builder) => builder
     ..messageBody = ''
     ..isRead = false;
 
-  static Query<Map<String, dynamic>> collection([DocumentReference parent]) =>
+  static Query<Map<String, dynamic>> collection([DocumentReference? parent]) =>
       parent != null
           ? parent.collection('Tour_Messages')
           : FirebaseFirestore.instance.collectionGroup('Tour_Messages');
@@ -47,11 +43,11 @@ abstract class TourMessagesRecord
 
   static Stream<TourMessagesRecord> getDocument(DocumentReference ref) => ref
       .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .map((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   static Future<TourMessagesRecord> getDocumentOnce(DocumentReference ref) =>
       ref.get().then(
-          (s) => serializers.deserializeWith(serializer, serializedData(s)));
+          (s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   TourMessagesRecord._();
   factory TourMessagesRecord(
@@ -61,19 +57,25 @@ abstract class TourMessagesRecord
   static TourMessagesRecord getDocumentFromData(
           Map<String, dynamic> data, DocumentReference reference) =>
       serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+          {...mapFromFirestore(data), kDocumentReferenceField: reference})!;
 }
 
 Map<String, dynamic> createTourMessagesRecordData({
-  DateTime dateSent,
-  DocumentReference senderUserReff,
-  String messageBody,
-  bool isRead,
-}) =>
-    serializers.toFirestore(
-        TourMessagesRecord.serializer,
-        TourMessagesRecord((t) => t
-          ..dateSent = dateSent
-          ..senderUserReff = senderUserReff
-          ..messageBody = messageBody
-          ..isRead = isRead));
+  DateTime? dateSent,
+  DocumentReference? senderUserReff,
+  String? messageBody,
+  bool? isRead,
+}) {
+  final firestoreData = serializers.toFirestore(
+    TourMessagesRecord.serializer,
+    TourMessagesRecord(
+      (t) => t
+        ..dateSent = dateSent
+        ..senderUserReff = senderUserReff
+        ..messageBody = messageBody
+        ..isRead = isRead,
+    ),
+  );
+
+  return firestoreData;
+}
