@@ -9,116 +9,88 @@ part 'tours_record.g.dart';
 abstract class ToursRecord implements Built<ToursRecord, ToursRecordBuilder> {
   static Serializer<ToursRecord> get serializer => _$toursRecordSerializer;
 
-  @nullable
   @BuiltValueField(wireName: 'created_time')
-  DateTime get createdTime;
+  DateTime? get createdTime;
 
-  @nullable
   @BuiltValueField(wireName: 'tour_name')
-  String get tourName;
+  String? get tourName;
 
-  @nullable
   @BuiltValueField(wireName: 'region_ID')
-  DocumentReference get regionID;
+  DocumentReference? get regionID;
 
-  @nullable
-  DocumentReference get uid;
+  DocumentReference? get uid;
 
-  @nullable
-  int get passengers;
+  int? get passengers;
 
-  @nullable
   @BuiltValueField(wireName: 'tour_date')
-  DateTime get tourDate;
+  DateTime? get tourDate;
 
-  @nullable
   @BuiltValueField(wireName: 'pickup_address')
-  String get pickupAddress;
+  String? get pickupAddress;
 
-  @nullable
   @BuiltValueField(wireName: 'pickup_latlng')
-  LatLng get pickupLatlng;
+  LatLng? get pickupLatlng;
 
-  @nullable
   @BuiltValueField(wireName: 'price_pp')
-  double get pricePp;
+  double? get pricePp;
 
-  @nullable
-  String get region;
+  String? get region;
 
-  @nullable
-  BuiltList<DocumentReference> get venues;
+  BuiltList<DocumentReference>? get venues;
 
-  @nullable
   @BuiltValueField(wireName: 'guests_uid')
-  BuiltList<DocumentReference> get guestsUid;
+  BuiltList<DocumentReference>? get guestsUid;
 
-  @nullable
   @BuiltValueField(wireName: 'pickup_unit_number')
-  String get pickupUnitNumber;
+  String? get pickupUnitNumber;
 
-  @nullable
   @BuiltValueField(wireName: 'tour_state')
-  String get tourState;
+  String? get tourState;
 
-  @nullable
-  double get platformTastingFee;
+  double? get platformTastingFee;
 
-  @nullable
   @BuiltValueField(wireName: 'driver_reff')
-  DocumentReference get driverReff;
+  DocumentReference? get driverReff;
 
-  @nullable
   @BuiltValueField(wireName: 'driver_uid')
-  String get driverUid;
+  String? get driverUid;
 
-  @nullable
   @BuiltValueField(wireName: 'country_state')
-  String get countryState;
+  String? get countryState;
 
-  @nullable
   @BuiltValueField(wireName: 'transport_fee_pp')
-  double get transportFeePp;
+  double? get transportFeePp;
 
-  @nullable
   @BuiltValueField(wireName: 'total_tasting_fee_pp')
-  int get totalTastingFeePp;
+  int? get totalTastingFeePp;
 
-  @nullable
   @BuiltValueField(wireName: 'large_group_venue_early_seating_count')
-  int get largeGroupVenueEarlySeatingCount;
+  int? get largeGroupVenueEarlySeatingCount;
 
-  @nullable
   @BuiltValueField(wireName: 'sub_total')
-  int get subTotal;
+  int? get subTotal;
 
-  @nullable
   @BuiltValueField(wireName: 'discount_amount')
-  int get discountAmount;
+  int? get discountAmount;
 
-  @nullable
   @BuiltValueField(wireName: 'total_paid')
-  int get totalPaid;
+  int? get totalPaid;
 
-  @nullable
   @BuiltValueField(wireName: 'total_balance')
-  int get totalBalance;
+  int? get totalBalance;
 
-  @nullable
   @BuiltValueField(wireName: 'promo_code')
-  String get promoCode;
+  String? get promoCode;
 
-  @nullable
   @BuiltValueField(wireName: 'promo_discount_percent')
-  double get promoDiscountPercent;
+  double? get promoDiscountPercent;
 
-  @nullable
   @BuiltValueField(wireName: 'promo_code_submitted')
-  String get promoCodeSubmitted;
+  String? get promoCodeSubmitted;
 
-  @nullable
   @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference get reference;
+  DocumentReference? get ffRef;
+  DocumentReference get reference => ffRef!;
 
   static void _initializeBuilder(ToursRecordBuilder builder) => builder
     ..tourName = ''
@@ -149,11 +121,11 @@ abstract class ToursRecord implements Built<ToursRecord, ToursRecordBuilder> {
 
   static Stream<ToursRecord> getDocument(DocumentReference ref) => ref
       .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .map((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   static Future<ToursRecord> getDocumentOnce(DocumentReference ref) => ref
       .get()
-      .then((s) => serializers.deserializeWith(serializer, serializedData(s)));
+      .then((s) => serializers.deserializeWith(serializer, serializedData(s))!);
 
   ToursRecord._();
   factory ToursRecord([void Function(ToursRecordBuilder) updates]) =
@@ -162,65 +134,71 @@ abstract class ToursRecord implements Built<ToursRecord, ToursRecordBuilder> {
   static ToursRecord getDocumentFromData(
           Map<String, dynamic> data, DocumentReference reference) =>
       serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference});
+          {...mapFromFirestore(data), kDocumentReferenceField: reference})!;
 }
 
 Map<String, dynamic> createToursRecordData({
-  DateTime createdTime,
-  String tourName,
-  DocumentReference regionID,
-  DocumentReference uid,
-  int passengers,
-  DateTime tourDate,
-  String pickupAddress,
-  LatLng pickupLatlng,
-  double pricePp,
-  String region,
-  String pickupUnitNumber,
-  String tourState,
-  double platformTastingFee,
-  DocumentReference driverReff,
-  String driverUid,
-  String countryState,
-  double transportFeePp,
-  int totalTastingFeePp,
-  int largeGroupVenueEarlySeatingCount,
-  int subTotal,
-  int discountAmount,
-  int totalPaid,
-  int totalBalance,
-  String promoCode,
-  double promoDiscountPercent,
-  String promoCodeSubmitted,
-}) =>
-    serializers.toFirestore(
-        ToursRecord.serializer,
-        ToursRecord((t) => t
-          ..createdTime = createdTime
-          ..tourName = tourName
-          ..regionID = regionID
-          ..uid = uid
-          ..passengers = passengers
-          ..tourDate = tourDate
-          ..pickupAddress = pickupAddress
-          ..pickupLatlng = pickupLatlng
-          ..pricePp = pricePp
-          ..region = region
-          ..venues = null
-          ..guestsUid = null
-          ..pickupUnitNumber = pickupUnitNumber
-          ..tourState = tourState
-          ..platformTastingFee = platformTastingFee
-          ..driverReff = driverReff
-          ..driverUid = driverUid
-          ..countryState = countryState
-          ..transportFeePp = transportFeePp
-          ..totalTastingFeePp = totalTastingFeePp
-          ..largeGroupVenueEarlySeatingCount = largeGroupVenueEarlySeatingCount
-          ..subTotal = subTotal
-          ..discountAmount = discountAmount
-          ..totalPaid = totalPaid
-          ..totalBalance = totalBalance
-          ..promoCode = promoCode
-          ..promoDiscountPercent = promoDiscountPercent
-          ..promoCodeSubmitted = promoCodeSubmitted));
+  DateTime? createdTime,
+  String? tourName,
+  DocumentReference? regionID,
+  DocumentReference? uid,
+  int? passengers,
+  DateTime? tourDate,
+  String? pickupAddress,
+  LatLng? pickupLatlng,
+  double? pricePp,
+  String? region,
+  String? pickupUnitNumber,
+  String? tourState,
+  double? platformTastingFee,
+  DocumentReference? driverReff,
+  String? driverUid,
+  String? countryState,
+  double? transportFeePp,
+  int? totalTastingFeePp,
+  int? largeGroupVenueEarlySeatingCount,
+  int? subTotal,
+  int? discountAmount,
+  int? totalPaid,
+  int? totalBalance,
+  String? promoCode,
+  double? promoDiscountPercent,
+  String? promoCodeSubmitted,
+}) {
+  final firestoreData = serializers.toFirestore(
+    ToursRecord.serializer,
+    ToursRecord(
+      (t) => t
+        ..createdTime = createdTime
+        ..tourName = tourName
+        ..regionID = regionID
+        ..uid = uid
+        ..passengers = passengers
+        ..tourDate = tourDate
+        ..pickupAddress = pickupAddress
+        ..pickupLatlng = pickupLatlng
+        ..pricePp = pricePp
+        ..region = region
+        ..venues = null
+        ..guestsUid = null
+        ..pickupUnitNumber = pickupUnitNumber
+        ..tourState = tourState
+        ..platformTastingFee = platformTastingFee
+        ..driverReff = driverReff
+        ..driverUid = driverUid
+        ..countryState = countryState
+        ..transportFeePp = transportFeePp
+        ..totalTastingFeePp = totalTastingFeePp
+        ..largeGroupVenueEarlySeatingCount = largeGroupVenueEarlySeatingCount
+        ..subTotal = subTotal
+        ..discountAmount = discountAmount
+        ..totalPaid = totalPaid
+        ..totalBalance = totalBalance
+        ..promoCode = promoCode
+        ..promoDiscountPercent = promoDiscountPercent
+        ..promoCodeSubmitted = promoCodeSubmitted,
+    ),
+  );
+
+  return firestoreData;
+}
