@@ -11,30 +11,30 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import '../../auth/auth_util.dart';
 
 String fortmatCurrency(
-  double price,
-  String currencySymbol,
+  double? price,
+  String? currencySymbol,
 ) {
   //return '$currencySymbol$price.00';
-  double num1 = double.parse((price).toStringAsFixed(2));
+  double num1 = double.parse(price!.toStringAsFixed(2));
   //return
   return '$currencySymbol$num1 pp';
 }
 
 DateTime getCurrentDateTimePlusAweek(
-  DateTime selectedDate,
-  int leadTime,
-  bool tourLeadTimeExempted,
+  DateTime? selectedDate,
+  int? leadTime,
+  bool? tourLeadTimeExempted,
 ) {
   var today = DateTime.now();
 
-  if (tourLeadTimeExempted) {
+  if (tourLeadTimeExempted!) {
     return DateTime(today.year, today.month, today.day + 1, 9);
   } else {
-    return DateTime(today.year, today.month, today.day + leadTime, 9);
+    return DateTime(today.year, today.month, today.day + leadTime!, 9);
   }
 }
 
-bool isStringNotEmpty(String aString) {
+bool isStringNotEmpty(String? aString) {
   return aString == null ? false : true;
 
   /*if (aString == "") {
@@ -45,20 +45,20 @@ bool isStringNotEmpty(String aString) {
 }
 
 LatLng createGeoPoint(
-  double latitude,
-  double longitude,
+  double? latitude,
+  double? longitude,
 ) {
   // Add your function code here!
-  return LatLng(latitude, longitude);
+  return LatLng(latitude!, longitude!);
 }
 
 bool isVenueAlreadyAdded(
-  List<SelectedVenuesRecord> data,
-  DocumentReference venue,
+  List<SelectedVenuesRecord>? data,
+  DocumentReference? venue,
 ) {
   // Add your function code here!
 
-  for (var i = 0; i < data.length; i++) {
+  for (var i = 0; i < data!.length; i++) {
     if (data[i].venueRef == venue) {
       return true;
     }
@@ -68,11 +68,11 @@ bool isVenueAlreadyAdded(
 }
 
 bool isVenueOpen(
-  List<int> openDays,
-  ToursRecord tourRecord,
+  List<int>? openDays,
+  ToursRecord? tourRecord,
 ) {
   // Add your function code here!
-  return openDays.contains(tourRecord.tourDate.weekday) ? true : false;
+  return openDays!.contains(tourRecord!.tourDate!.weekday) ? true : false;
 
   /*for (var i = 0; i < openDays.length; i++) {
     if (openDays[i] == tourRecord.tourDate.weekday) {
@@ -84,13 +84,13 @@ bool isVenueOpen(
 }
 
 bool isVenueClosedOrExceedsCapacity(
-  List<int> openDays,
-  ToursRecord tourRecord,
-  int venueCapacity,
+  List<int>? openDays,
+  ToursRecord? tourRecord,
+  int? venueCapacity,
 ) {
   // Add your function code here!
-  return !openDays.contains(tourRecord.tourDate.weekday) ||
-          tourRecord.passengers > venueCapacity
+  return !openDays!.contains(tourRecord!.tourDate!.weekday) ||
+          tourRecord.passengers! > venueCapacity!
       ? true
       : false;
 
@@ -104,47 +104,47 @@ bool isVenueClosedOrExceedsCapacity(
 }
 
 bool doesTourExceedsVenueCapacity(
-  ToursRecord tourRecord,
-  int venueCapacity,
+  ToursRecord? tourRecord,
+  int? venueCapacity,
 ) {
   //return tourRecord != null && venueCapacity != null && tourRecord.passengers! > venueCapacity ? true : false;
 
-  return tourRecord.passengers > venueCapacity ? true : false;
+  return tourRecord!.passengers! > venueCapacity! ? true : false;
 }
 
 bool meetsVenueCapacity(
-  ToursRecord tourRecord,
-  int venueCapacity,
+  ToursRecord? tourRecord,
+  int? venueCapacity,
 ) {
   // Add your function code here!
-  return tourRecord.passengers <= venueCapacity ? true : false;
+  return tourRecord!.passengers! <= venueCapacity! ? true : false;
 }
 
-String upperCaseString(String str) {
+String upperCaseString(String? str) {
   // Add your function code here!
-  return str.toUpperCase();
+  return str!.toUpperCase();
 }
 
-String getDayMonthFromDate(DateTime date) {
+String getDayMonthFromDate(DateTime? date) {
   // Get month from Date
-  return DateFormat("d, MMMM").format(date);
+  return DateFormat("d, MMMM").format(date!);
 }
 
 String concatenateStrings(
-  String string1,
-  String string2,
+  String? string1,
+  String? string2,
 ) {
   // Concatenate two strings
-  if (string1.isNotEmpty && string2.isNotEmpty) {
+  if (string1!.isNotEmpty && string2!.isNotEmpty) {
     return '$string1 $string2';
   }
-  return string1.isNotEmpty ? string1 : string2;
+  return string1.isNotEmpty ? string1 : string2!;
 }
 
-String getTimeFromDate(DateTime date) {
+String getTimeFromDate(DateTime? date) {
   // get time from date
 
-  if (date.year == 1970) {
+  if (date!.year == 1970) {
     return "TBC";
   } else {
     return DateFormat.jm().format(date);
@@ -156,8 +156,8 @@ String getTimeFromDate(DateTime date) {
 }
 
 String getLatestDateDayMonthValue(
-  DateTime originalTourDate,
-  DateTime editedTourDate,
+  DateTime? originalTourDate,
+  DateTime? editedTourDate,
 ) {
   // check if editedTourDate is not empty and is greater than originalTourDate
   if (editedTourDate == null) {
@@ -175,7 +175,7 @@ String getLatestDateDayMonthValue(
   }
 }
 
-bool isArrayEmpty(List<DocumentReference> arrayList) {
+bool isArrayEmpty(List<DocumentReference>? arrayList) {
   // Check if arrayList is empty then return true else false
   if (arrayList == null)
     return false;
@@ -184,23 +184,23 @@ bool isArrayEmpty(List<DocumentReference> arrayList) {
 }
 
 bool isTourStopsCountLessThanLimitAllowed(
-  List<DocumentReference> venuesList,
-  AppConfigRecord venueLimit,
+  List<DocumentReference>? venuesList,
+  AppConfigRecord? venueLimit,
 ) {
   // Check if venuesList count is less than 4, if true return true else false
-  return venuesList.length < venueLimit.itineraryVenueLimit ? true : false;
+  return venuesList!.length < venueLimit!.itineraryVenueLimit! ? true : false;
 }
 
-String boolString(bool boolVal) {
+String boolString(bool? boolVal) {
   // Add your function code here!
-  if (boolVal) {
+  if (boolVal!) {
     return 'true';
   } else {
     return 'false';
   }
 }
 
-bool isTourInDraftState(String stateString) {
+bool isTourInDraftState(String? stateString) {
   // Add your function code here!
 
   return stateString == "draft" ? true : false;
@@ -212,16 +212,18 @@ DateTime epochTime() {
 }
 
 String getPerPersonFee(
-  double transportFee,
-  List<SelectedVenuesRecord> selectedVenues,
-  double platformFee,
+  double? transportFee,
+  List<SelectedVenuesRecord>? selectedVenues,
+  double? platformFee,
 ) {
   //final currencyFormatter = NumberFormat.currency();
-  double totalPerPerson = transportFee;
+  double? totalPerPerson = transportFee;
 
-  for (var i = 0; i < selectedVenues.length; i++) {
-    totalPerPerson += selectedVenues[i].tastingFee + platformFee;
-    print(totalPerPerson);
+  for (var i = 0; i < selectedVenues!.length; i++) {
+    if (totalPerPerson != null) {
+      totalPerPerson += selectedVenues[i].tastingFee! + platformFee!;
+      print(totalPerPerson);
+    }
   }
 
   return NumberFormat.simpleCurrency().format(totalPerPerson);
@@ -230,15 +232,30 @@ String getPerPersonFee(
 }
 
 int getPerPersonFeeAsInt(
-  double transportFee,
-  List<SelectedVenuesRecord> selectedVenues,
-  double platformFee,
+  double? transportFee,
+  List<SelectedVenuesRecord>? selectedVenues,
+  double? platformFee,
 ) {
   //final currencyFormatter = NumberFormat.currency();
+  if (transportFee == null) {
+    throw ArgumentError(
+        "You should enter a valid transportFee for transportFee, not null or empty");
+  }
+
+  if (selectedVenues == null) {
+    throw ArgumentError(
+        "You should enter a valid selectedVenues for selectedVenues, not null or empty");
+  }
+
+  if (platformFee == null) {
+    throw ArgumentError(
+        "You should enter a valid platformFee for platformFee, not null or empty");
+  }
+
   double totalPerPerson = transportFee;
 
   for (var i = 0; i < selectedVenues.length; i++) {
-    totalPerPerson += selectedVenues[i].tastingFee + platformFee;
+    totalPerPerson += selectedVenues[i].tastingFee! + platformFee;
     print(totalPerPerson);
   }
 
@@ -248,30 +265,34 @@ int getPerPersonFeeAsInt(
 }
 
 int getPerPersonFeeAsInt2(
-  double transportFee,
-  List<SelectedVenuesRecord> selectedVenues,
-  double platformFee,
-  SelectedVenuesRecord lastSelectedVenue,
+  double? transportFee,
+  List<SelectedVenuesRecord>? selectedVenues,
+  double? platformFee,
+  SelectedVenuesRecord? lastSelectedVenue,
 ) {
-  //final currencyFormatter = NumberFormat.currency();
-  double totalPerPerson = transportFee;
-  final List<SelectedVenuesRecord> venues = selectedVenues;
+  double? totalPerPerson = transportFee;
+  final List<SelectedVenuesRecord> venues = selectedVenues!;
 
-  venues.add(lastSelectedVenue);
+  if (lastSelectedVenue != null) {
+    venues.add(lastSelectedVenue);
+  }
 
   for (var i = 0; i < venues.length; i++) {
-    totalPerPerson += venues[i].tastingFee + platformFee;
-    print(totalPerPerson);
+    if (totalPerPerson != null) {
+      totalPerPerson += venues[i].tastingFee! + platformFee!;
+      print(totalPerPerson);
+    }
   }
-  return totalPerPerson.toInt();
+
+  return totalPerPerson!.toInt();
 }
 
 DateTime getBookingReservationTime(
-  DateTime tourDate,
-  DateTime reservationTimestamp,
+  DateTime? tourDate,
+  DateTime? reservationTimestamp,
 ) {
-  return DateTime(tourDate.year, tourDate.month, tourDate.day,
-      reservationTimestamp.hour, reservationTimestamp.minute);
+  return DateTime(tourDate!.year, tourDate.month, tourDate.day,
+      reservationTimestamp!.hour, reservationTimestamp.minute);
 }
 
 DateTime getTodayTimestampZeroMinutes() {
@@ -279,101 +300,126 @@ DateTime getTodayTimestampZeroMinutes() {
       DateTime.now().hour, 0);
 }
 
-int countNumberOfGuests(List<DocumentReference> guestsList) {
+int countNumberOfGuests(List<DocumentReference>? guestsList) {
   // Add your function code here!
-  return guestsList.length;
+  return guestsList!.length;
 }
 
 bool isLunchVenue(
-  DocumentReference selectedVenueReff,
-  DocumentReference localStateSelectedVenueReff,
+  DocumentReference? selectedVenueReff,
+  DocumentReference? localStateSelectedVenueReff,
 ) {
   // Add your function code here!
   return selectedVenueReff == localStateSelectedVenueReff ? true : false;
 }
 
-bool isStringEmpty(String aString) {
+bool isStringEmpty(String? aString) {
   // Add your function code here!
-  return aString.isEmpty;
+  return aString!.isEmpty;
 }
 
 int calculateTotalTastingFeePP(
-  List<SelectedVenuesRecord> selectedVenues,
-  double selectedVenueFee,
+  List<SelectedVenuesRecord>? selectedVenues,
+  double? selectedVenueFee,
 ) {
   // Add your function code here!
   int totalTastingFeePP = 0;
 
-  for (var i = 0; i < selectedVenues.length; i++) {
-    totalTastingFeePP += selectedVenues[i].tastingFee.toInt();
+  for (var i = 0; i < selectedVenues!.length; i++) {
+    totalTastingFeePP += selectedVenues[i].tastingFee!.toInt();
     print(totalTastingFeePP);
   }
 
-  return totalTastingFeePP + selectedVenueFee.toInt();
+  return totalTastingFeePP + selectedVenueFee!.toInt();
 }
 
 int deductFromTotalTastingFeePP(
-  int currentTastingFee,
-  SelectedVenuesRecord subtractValue,
+  int? currentTastingFee,
+  SelectedVenuesRecord? subtractValue,
 ) {
   // Add your function code here!
+  if (currentTastingFee == null) {
+    throw ArgumentError(
+        "You should enter a valid currentTastingFee for editedTourDate, not null or empty");
+  }
+
+  if (subtractValue == null) {
+    throw ArgumentError(
+        "You should enter a valid currentTastingFee for editedTourDate, not null or empty");
+  }
+
   return currentTastingFee -
-      subtractValue.tastingFee.toInt(); //subtractValue.toInt();
+      subtractValue.tastingFee!.toInt(); //subtractValue.toInt();
 }
 
 int addtoTotalTastingFeePP(
-  int currentTastingFee,
-  double addValue,
+  int? currentTastingFee,
+  double? addValue,
 ) {
   // Add your function code here!
-  return currentTastingFee + addValue.toInt();
+  return currentTastingFee! + addValue!.toInt();
 }
 
 String setImagePath(
-  String tastingExperienceRecordImgPath,
-  String venueImgPath,
+  String? tastingExperienceRecordImgPath,
+  String? venueImgPath,
 ) {
   // Add your function code here!
-  return tastingExperienceRecordImgPath.isEmpty
-      ? venueImgPath
+  return tastingExperienceRecordImgPath!.isEmpty
+      ? venueImgPath!
       : tastingExperienceRecordImgPath;
 }
 
 bool isIntegerOneSmallOrEqualThanIntegerTwo(
-  int intergerOne,
-  int intergerTwo,
+  int? intergerOne,
+  int? intergerTwo,
 ) {
   // Add your function code here!
-  return intergerOne <= intergerTwo ? true : false;
+  return intergerOne! <= intergerTwo! ? true : false;
 }
 
 bool isIntegerOneSmallerThanIntegerTwoOrArguement3IsTrue(
-  int intergerOne,
-  int intergerTwo,
-  bool arguement3,
+  int? intergerOne,
+  int? intergerTwo,
+  bool? arguement3,
 ) {
   // Add your function code here!
-  return intergerOne < intergerTwo || !arguement3 ? true : false;
+  return intergerOne! < intergerTwo! || !arguement3! ? true : false;
 }
 
-String displayTastingExperienceDescription(String tastingDescription) {
+String displayTastingExperienceDescription(String? tastingDescription) {
   // Add your function code here!
-  return tastingDescription == "" ? "Not applicable" : tastingDescription;
+  return tastingDescription == "" ? "Not applicable" : tastingDescription!;
 }
 
 int getTourSubTotal(
-  int noOfPassengers,
-  double costPP,
+  int? noOfPassengers,
+  double? costPP,
 ) {
   // Add your function code here!
+  if (noOfPassengers == null) {
+    throw ArgumentError(
+        "You should enter a valid noOfPassengers for noOfPassengers, not null or empty");
+  }
+
+  if (costPP == null) {
+    throw ArgumentError(
+        "You should enter a valid costPP for costPP, not null or empty");
+  }
+
   //int costpp = (costPP / 100).round();
   return noOfPassengers * costPP.toInt();
 }
 
-bool isBoolTrue(bool value) {
+bool isBoolTrue(bool? value) {
   // Add your function code here!
   if (value.toString() != "") {
-    return value;
+    return value!;
   }
   return false;
+}
+
+String tourRefAsString(DocumentReference? tourReff) {
+  // Add your function code here!
+  return tourReff.toString();
 }
