@@ -23,12 +23,12 @@ enum MediaSource {
   camera,
 }
 
-Future<List<SelectedMedia>> selectMediaWithSourceBottomSheet({
-  BuildContext context,
-  double maxWidth,
-  double maxHeight,
-  int imageQuality,
-  bool allowPhoto,
+Future<List<SelectedMedia>?> selectMediaWithSourceBottomSheet({
+  required BuildContext context,
+  double? maxWidth,
+  double? maxHeight,
+  int? imageQuality,
+  required bool allowPhoto,
   bool allowVideo = false,
   String pickerFontFamily = 'Roboto',
   Color textColor = const Color(0xFF111417),
@@ -118,10 +118,10 @@ Future<List<SelectedMedia>> selectMediaWithSourceBottomSheet({
   );
 }
 
-Future<List<SelectedMedia>> selectMedia({
-  double maxWidth,
-  double maxHeight,
-  int imageQuality,
+Future<List<SelectedMedia>?> selectMedia({
+  double? maxWidth,
+  double? maxHeight,
+  int? imageQuality,
   bool isVideo = false,
   MediaSource mediaSource = MediaSource.camera,
   bool multiImage = false,
@@ -163,7 +163,7 @@ Future<List<SelectedMedia>> selectMedia({
   if (mediaBytes == null) {
     return null;
   }
-  final path = storagePath(currentUserUid, pickedMedia.name, isVideo);
+  final path = storagePath(currentUserUid, pickedMedia!.name, isVideo);
   return [SelectedMedia(path, mediaBytes)];
 }
 
@@ -179,7 +179,7 @@ bool validateFileFormat(String filePath, BuildContext context) {
   return false;
 }
 
-Future<SelectedMedia> selectFile({
+Future<SelectedMedia?> selectFile({
   List<String> allowedExtensions = const ['pdf'],
 }) async {
   final pickedFiles = await FilePicker.platform.pickFiles(
@@ -196,10 +196,10 @@ Future<SelectedMedia> selectFile({
     return null;
   }
   final path = storagePath(currentUserUid, file.name, false);
-  return SelectedMedia(path, file.bytes);
+  return SelectedMedia(path, file.bytes!);
 }
 
-String storagePath(String uid, String filePath, bool isVideo, [int index]) {
+String storagePath(String uid, String filePath, bool isVideo, [int? index]) {
   final timestamp = DateTime.now().microsecondsSinceEpoch;
   // Workaround fixed by https://github.com/flutter/plugins/pull/3685
   // (not yet in stable).
